feat: app 클래스 분리
